refactor: use g-c-cpp-common's async retry loop
