+++ google/cloud/spanner/internal/session_pool.cc
@@ -13,11 +13,11 @@
 // limitations under the License.
 
 #include "google/cloud/spanner/internal/session_pool.h"
-#include "google/cloud/spanner/internal/async_retry_unary_rpc.h"
 #include "google/cloud/spanner/internal/connection_impl.h"
 #include "google/cloud/spanner/internal/retry_loop.h"
 #include "google/cloud/spanner/internal/session.h"
 #include "google/cloud/completion_queue.h"
+#include "google/cloud/internal/async_retry_unary_rpc.h"
 #include "google/cloud/internal/make_unique.h"
 #include "google/cloud/log.h"
 #include "google/cloud/status.h"
@@ -417,16 +417,16 @@ SessionPool::AsyncBatchCreateSessions(
   request.mutable_session_template()->mutable_labels()->insert(labels.begin(),
                                                                labels.end());
   request.set_session_count(std::int32_t{num_sessions});
-  return internal::StartRetryAsyncUnaryRpc(
-      __func__, retry_policy_prototype_->clone(),
+  return google::cloud::internal::StartRetryAsyncUnaryRpc(
+      cq, __func__, retry_policy_prototype_->clone(),
       backoff_policy_prototype_->clone(),
-      internal::ConstantIdempotencyPolicy(true),
+      /*is_idempotent=*/true,
       [stub](grpc::ClientContext* context,
              spanner_proto::BatchCreateSessionsRequest const& request,
              grpc::CompletionQueue* cq) {
         return stub->AsyncBatchCreateSessions(*context, request, cq);
       },
-      std::move(request), cq);
+      std::move(request));
 }
 
 future<StatusOr<google::protobuf::Empty>> SessionPool::AsyncDeleteSession(
@@ -435,16 +435,16 @@ future<StatusOr<google::protobuf::Empty>> SessionPool::AsyncDeleteSession(
     std::shared_ptr<SpannerStub> stub, std::string session_name) {
   spanner_proto::DeleteSessionRequest request;
   request.set_name(std::move(session_name));
-  return internal::StartRetryAsyncUnaryRpc(
-      __func__, retry_policy_prototype_->clone(),
+  return google::cloud::internal::StartRetryAsyncUnaryRpc(
+      cq, __func__, retry_policy_prototype_->clone(),
       backoff_policy_prototype_->clone(),
-      internal::ConstantIdempotencyPolicy(true),
+      /*is_idempotent=*/true,
       [stub](grpc::ClientContext* context,
              spanner_proto::DeleteSessionRequest const& request,
              grpc::CompletionQueue* cq) {
         return stub->AsyncDeleteSession(*context, request, cq);
       },
-      std::move(request), cq);
+      std::move(request));
 }
 
 future<StatusOr<spanner_proto::Session>> SessionPool::AsyncGetSession(
@@ -453,16 +453,16 @@ future<StatusOr<spanner_proto::Session>> SessionPool::AsyncGetSession(
     std::shared_ptr<SpannerStub> stub, std::string session_name) {
   spanner_proto::GetSessionRequest request;
   request.set_name(std::move(session_name));
-  return internal::StartRetryAsyncUnaryRpc(
-      __func__, retry_policy_prototype_->clone(),
+  return google::cloud::internal::StartRetryAsyncUnaryRpc(
+      cq, __func__, retry_policy_prototype_->clone(),
       backoff_policy_prototype_->clone(),
-      internal::ConstantIdempotencyPolicy(true),
+      /*is_idempotent=*/true,
       [stub](grpc::ClientContext* context,
              spanner_proto::GetSessionRequest const& request,
              grpc::CompletionQueue* cq) {
         return stub->AsyncGetSession(*context, request, cq);
       },
-      std::move(request), cq);
+      std::move(request));
 }
 
 Status SessionPool::HandleBatchCreateSessionsDone(
