+++ google/cloud/spanner/spanner_client.bzl
@@ -36,8 +36,6 @@ spanner_client_hdrs = [
     "instance_admin_client.h",
     "instance_admin_connection.h",
     "internal/api_client_header.h",
-    "internal/async_retry_op.h",
-    "internal/async_retry_unary_rpc.h",
     "internal/build_info.h",
     "internal/channel.h",
     "internal/compiler_info.h",
