+++ google/cloud/spanner/spanner_client_testing.bzl
@@ -19,11 +19,9 @@
 spanner_client_testing_hdrs = [
     "testing/database_environment.h",
     "testing/matchers.h",
-    "testing/mock_completion_queue.h",
     "testing/mock_database_admin_stub.h",
     "testing/mock_instance_admin_stub.h",
     "testing/mock_partial_result_set_reader.h",
-    "testing/mock_response_reader.h",
     "testing/mock_spanner_stub.h",
     "testing/pick_instance_config.h",
     "testing/pick_random_instance.h",
